perf(valdres): keep equality count on subscriber set
